+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [v0.2.1] - 2026-06-
+### Added
+- Remove offline servers from `sing-box` config (for types "selector" and "urltest")
+- Keep connection alive and switch to another server immediately if not
+
 ## [v0.1.5] - 2026-06-02
 ### Added
 - Bug fixes for Trojan